etl/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
Django==3.2
django-cors-headers==2.2.0
django-split-settings==1.2.0
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==8.8.0
flake8==4.0.0
gunicorn==20.0.4
mccabe==0.6.1
psycopg2-binary==2.9.6
pycodestyle==2.8.0
pydantic==1.10.9
pydentic==0.0.1.dev3
pyflakes==2.4.0
python-dotenv==0.21.0
python-stdnum==1.18
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==1.26.16
redis==4.4.2